fix: 당첨율 소수점 잘리는 오류 수정

+++ src/main/java/dev/dahye/lotto/domain/LottoResult.java
@@ -103,6 +103,6 @@ public class LottoResult {
     }
 
     public static double getWinningRate(int totalPrize, int money) {
-        return DoubleUtils.parseDoubleSecondDigit(totalPrize / money);
+        return DoubleUtils.parseDoubleSecondDigit((double) totalPrize / money);
     }
 }
